Proposed edits
As discussed in email, 

This softens the somewhat pointed recommendation about objecting on specific issues rather than on a charter.  That used the word "proxy", which might be interpreted as being overly negative.

It removes the recommendation about the order in which the group does its work.  The order suggested has a spec published with unaddressed errata, which is a fair choice, but it isn't really necessary to make such a specific recommendation, especially when this ordering point has already been made (and better, in my view).